fix: template selecionado no drawer não chegava ao editor
O fluxo de criação navega via URL params — as seções do template eram
construídas no cliente mas descartadas na navegação. Agora:
- ServicePlanConfig inclui templateId
- ClientServicesPage passa templateId na URL
- new/page.tsx busca o template no servidor e monta as seções
- ClientServiceFormPage aceita defaultMeta.sections e repassa ao editor
- buildSectionsFromTemplate extraído para lib/template-utils.ts (compartilhado
  entre cliente e servidor)
